+++ examples/index.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Example from './basic';
+
+ReactDOM.render(<Example />, document.getElementById('root'));