+++ src/useConstraints.ts
@@ -1,10 +1,11 @@
 import { useMemo } from "react";
-import { Values, MutableValues, ConstantValues } from "./useResetableValues";
+import { Values, MutableValues, ConstantValues } from "./useGenericValues";
 import {
-    Validator, Validators, SingleValidator, validateValidators, userSuppliedValue,
+    Validator, Validators, SingleValidator, validateValidators,
 } from "./useValidation";
 import { assert, LoggingTypes } from "./utils";
 import { Errors } from "./useErrors";
+import { userSuppliedValue, Fields } from "./useFields";
 
 type supportedTypes = "email" | "text" | "url" | "number" | "date";
 // possible values:
@@ -24,11 +25,11 @@ interface Constraints {
     /**
      * A min boundary used for type numbers
      */
-    readonly min?: number | Date | [number | Date, string];
+    readonly min?: number | string | [number | string, string];
     /**
      * A max boundary used for type numbers
      */
-    readonly max?: number | Date | [number | Date, string];
+    readonly max?: number | string | [number | string, string];
     /**
      * Determines if the field is required
      *
@@ -57,7 +58,7 @@ type RequiredConstraint<T extends supportedConstraints> = {
 export type ConstraintValidators = Values<Constraints | Validator>;
 
 interface SyncedConstraint {
-    (values: Values<userSuppliedValue>): ConstraintValidators;
+    (values: Fields): ConstraintValidators;
 }
 
 const defaultMessage = {
@@ -148,9 +149,10 @@ const propertyValidators = {
             regex = /^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$/;
             return regex.test(value);
         case "number":
-            // could use: /^-?(\d+|\d+\.\d+|\.\d+)([eE][-+]?\d+)?$/
-            return !Number.isNaN(Number(value));
+            return /^-?(\d+|\d+\.\d+|\.\d+)([eE][-+]?\d+)?$/.test(value);
         case "date":
+            // TODO: update to date yyyy-MM-dd
+            // TODO: support datetime-local and datetime with warning
             return !Number.isNaN((new Date(value)).getTime());
         case "text":
         default: return true;
@@ -313,7 +315,7 @@ export function useConstraints(
 ): Validators | SingleValidator<userSuppliedValue> {
     return useMemo((): Validators | SingleValidator<userSuppliedValue> => {
         if (typeof rules === "function") {
-            return (values: Values<userSuppliedValue>): Promise<Errors> => {
+            return (values: Fields): Promise<Errors> => {
                 const constraints = rules(values);
                 const validators = mapConstraintsToValidators(constraints);
                 const names = Object.keys(constraints);