+++ src/utils.ts
@@ -1,4 +1,4 @@
-const projectName = "JustAnotherValidator";
+const projectName = "react-uniformed";
 
 export enum LoggingTypes {
     invalidArgument = "InvalidArgument",
@@ -6,41 +6,22 @@ export enum LoggingTypes {
     constraintError = "ConstraintError",
 }
 interface Logger {
-    (type: LoggingTypes, message: string): void;
+    (message?: string, ...optionalParams: string[]): void;
 }
 
-interface Log {
-    readonly [key: string]: Logger;
-}
-
-/* eslint-disable no-console */
 // eslint-disable-next-line import/prefer-default-export
-const log: Log = {
-    warning(type: LoggingTypes, message: string): void {
-        console.warn(`${projectName}: [${type}] ${message}`);
-    },
-    debug(type: LoggingTypes, message: string): void {
-        console.debug(`${projectName}: [${type}] ${message}`);
-    },
-    error(type: LoggingTypes, message: string): void {
-        console.error(`${projectName}: [${type}] ${message}`);
-    },
-};
-
+/* eslint-disable no-console */
 function assertion(condition: boolean, type: LoggingTypes, message: string, logger: Logger): void {
     if (!condition) {
-        logger(type, message);
+        logger(`${projectName}: [${type}] ${message}`);
     }
 }
 
 export const assert = {
     error(condition: boolean, type: LoggingTypes, message: string): void {
-        assertion(condition, type, message, log.error);
+        assertion(condition, type, message, console.error);
     },
     warning(condition: boolean, type: LoggingTypes, message: string): void {
-        assertion(condition, type, message, log.warning);
-    },
-    debug(condition: boolean, type: LoggingTypes, message: string): void {
-        assertion(condition, type, message, log.debug);
+        assertion(condition, type, message, console.warn);
     },
 };