+++ webpack.examples.config.js
@@ -1,0 +1,34 @@
+const path = require("path");
+const webpack = require("webpack");
+
+module.exports = {
+    entry: ["babel-polyfill", "./examples/index.tsx"],
+    mode: "development",
+    module: {
+        rules: [
+            {
+                test: /\.tsx?$/,
+                exclude: /(node_modules|bower_components)/,
+                loader: "babel-loader",
+                options: {
+                    presets: ["@babel/env", "@babel/react", "@babel/typescript"],
+                },
+            },
+        ],
+    },
+    resolve: {
+        extensions: [".tsx", ".ts", ".js"],
+    },
+    output: {
+        path: path.resolve(__dirname, "dist/"),
+        publicPath: "/dist/",
+        filename: "bundle.js",
+    },
+    devServer: {
+        contentBase: path.join(__dirname, "dist/"),
+        port: 3000,
+        publicPath: "http://localhost:3000/examples/",
+        hotOnly: true,
+    },
+    plugins: [new webpack.HotModuleReplacementPlugin()],
+};